upload/requirements.txt
Flask==2.2.2
flask_sqlalchemy==3.0.3
python-dotenv==0.21.1
requests==2.28.2
